requirements.txt
Django==5.2.6
djangorestframework==3.15.1
psycopg2-binary==2.9.9
django-redis==5.4.0
python-dotenv==1.0.1
google-genai==1.70.0